% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{BINCO.genie3}
\alias{BINCO.genie3}
\title{Run the GENIE3 tree-based variable selection procedure on B bootstrap samples.}
\usage{
BINCO.genie3(Y, X, nb, prop_select = 0.2)
}
\arguments{
\item{Y}{target variables}

\item{X}{features (covariates)}

\item{nb}{Number of bootstrap iterations}

\item{prop_select}{What proportion of variables to select at each bootstrap iteration.}
}
\description{
Run the GENIE3 tree-based variable selection procedure on B bootstrap samples.
}
