% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{BINCO.qvals}
\alias{BINCO.qvals}
\title{Given selection frequencies, use BINCO to compute q-values.}
\usage{
BINCO.qvals(selection_frequency, nb)
}
\arguments{
\item{selection_frequency}{Selection frequencies: one non-negative integer per variable in your regression.}

\item{nb}{Number of bootstrap resamples (max possible value of selection_frequency)}
}
\value{
A vector of q-values with the same length as selection_frequency.
}
\description{
Given selection frequencies, use BINCO to compute q-values.
}
