\name{BINCOSimulation}
\alias{BINCOSimulation}
\title{A simulated data example for \code{BINCO}}
\description{A list containing a simulated data example for package \code{BINCO}}
\value{
BINCOSimulation is a list containing three components:
\item{Y}{a numerical vector of length=100.} 
\item{Y.m}{a numerical matrix consisting of 500 rows and 500 columns.}
\item{nb}{a positive integer.}
}

\details{
\code{Y} is generated by applying space on \code{B=100} bootstrap resamples from a simulated dataset based on a powerlaw network.

\code{Y.m} is the selection frequencies matrix.

\code{nb} is the number of (bootstrap) resamples used to generate edge selection frequencies.
}
 
\references{S. Li, L. Hsu, J. Peng and P. Wang (2012) Bootstrap Inference for Network Construction.
            }
            
\usage{data(BINCOSimulation)}

\keyword{datasets}

