% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{bio_beta_power}
\alias{bio_beta_power}
\title{1. kernel of the convoluted powered beta +binomial density #}
\usage{
bio_beta_power(n, k, a, b, r, x)
}
\arguments{
\item{n}{@param k parameters for binomial density}

\item{a}{@param b @param r parameters for the powered beta}
}
\description{
1. kernel of the convoluted powered beta +binomial density #
}
