% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{bio_beta_power_int}
\alias{bio_beta_power_int}
\title{2. numerical integration of f(x=k/n) #}
\usage{
bio_beta_power_int(n, k, a, b, r)
}
\arguments{
\item{n}{@param k parameters for binomial density}

\item{a}{@param b @param r parameters for the powered beta}
}
\description{
2. numerical integration of f(x=k/n) #
}
