% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{null.estimate}
\alias{null.estimate}
\title{8. estimate the null}
\usage{
null.estimate(count.mix, mu, vp, niv, mod)
}
\arguments{
\item{count.mix}{is the empirical density to fit}

\item{vp}{is the valley point value}
}
\description{
8. estimate the null
}
