% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{pick.close}
\alias{pick.close}
\title{4. find the index of the entry in a vector #
   that is closest to some value c         #}
\usage{
pick.close(v, c)
}
\arguments{
\item{v}{a vector}

\item{c}{a real number}
}
\description{
4. find the index of the entry in a vector #
   that is closest to some value c         #
}
