% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{u_shape_test}
\alias{u_shape_test}
\title{7. u-shape detection via three criteria.}
\usage{
u_shape_test(count.mix, vpr = 0.8)
}
\description{
The criteria:
}
\details{
1: valley point < threshold?
2: Decreasing at the beginning?
3: increasing at the tail?
}
