% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BINCO.r
\name{BINCO}
\alias{BINCO}
\title{Run the BINCO procedure}
\usage{
BINCO(
  count.mix = NULL,
  freq.m = NULL,
  nb = NULL,
  FDR = 0.05,
  vpr = 0.8,
  conservative = F,
  niv = 3,
  ini.bound = c(10, 10, 20)
)
}
\arguments{
\item{count.mix}{is a vector of non-negative integers as the counts of edges of 
different selection frequencies. Its length is the number of (bootstrap) 
resamples from which is was generated. For example, if the second entry is 50 
then it means 50 edges/variables are selected exactly twice while applying the 
original procedure on all resamples. The last entry is the number of 
edges/variables that are selected consistently while applying the original 
procedure on all resamples.}

\item{freq.m}{is a matrix recording the selection frequencies of all edges, e.g.,
freq.m[i,j] is the selection frequency of the edge connecting nodes i and j. 
It must be provided if count.mix is not.}

\item{nb}{is a positive integer as the number of (bootstrap) resmaples used to generate
count.mix of freq.m. It must be provided along with freq.m if count.mix is not
provided.}

\item{FDR}{is a numeric value between 0 and 1, the desired control level for the false
discovery rate. Default is 0.05.}

\item{vpr}{is the maximum allowed value for the "valley point"
of the empirical distribution (see Li, et al.,2012). Default value is 0.8. 
Large valley point values may result in liberal selection: above 0.95 is not recommended.}

\item{conservative}{is a logic value of TRUE or FALSE. Default is FALSE. Set 
conservative=TRUE if a conservative selection is needed (i.e., when the FDR 
control needs to be stringent). We recommand to apply \code{BINCO} under the 
conservative mode for the data where its "valley point" value is large.}

\item{niv}{is a positive integer as the number of sets of initial parameter values to
be used for density fitting. Default is 3, and 10 is large enough for most 
usual situations.}

\item{ini.bound}{is a vector of three positive real numbers. It gives the upper bound
of randomly generated initial parameter values used for density fitting. Defualt
value is (10,10,20).}
}
\description{
Run the BINCO procedure
}
